Added test support with an inmemory room and mocked server response
Just meant adding all new test components and modules, and setting the
runner to call the mock application that implements the test component.
Right now, the ImageApi returns a fixed thing that it mocks and it
doesn't return anything else. The could perhaps be improved by moving
the ImageApi provider to the application level, injecting it into the
test itself and setting the 'when' clauses there.