Almost everything that's needed for the DaggerMocks to work
The only issue is that the ModuleClasses can't be mocked for some
reason.
mock-maker-inline is no longer necessary as of mockito 2.7.5 or
something and definitely mockito 3 doesn't need it, which is the one
we're using right now.
So if you go and open all module classes (there's like 3) then this code
works!
However the other issue is that I don't know how to use a mix of
concrete classes with the mocks, and that's what's needed to really test
paging as it's supposed to work.
The room on disk needs to be replaced with a room memory database or we
need a way to clear it out everytime.
Maybe with a test rule for that. This might actually be preferable to a
room that stores data in memory, when we have a massive amount of data
that's needed to test the paging library.
androidTest mockito lib can be removed because daggermock contains it.